+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.2.0
+
+- Solved major integration bug (github.com/dentemple/jest-invert/issues/1)
+- Removed workaround (now defunct) caused by aforementioned bug
+- Renamed the `.isActive` method to `.run`
+- Refactored main
+- Refactored main tests
+- Added integration tests
+- Converted to ES6
+- Arrays and Objects are now part of the main API; removed the now defunct `.all()` API option
+- Updated README
+- Set `jest` as a peer dependency
+
 ## v0.1.3
 
 - Minor updates to docs